Add unit tests.
