+++ AdvancedDLSupport.Tests/Data/Interfaces/INullableLibrary.cs
@@ -1,0 +1,15 @@
+﻿namespace AdvancedDLSupport.Tests.Data
+{
+    public interface INullableLibrary
+    {
+        TestStruct? GetAllocatedTestStruct();
+        TestStruct? GetNullTestStruct();
+        bool CheckIfStructIsNull(TestStruct? testStruct);
+
+        [NativeSymbol(nameof(CheckIfStructIsNull))]
+        bool CheckIfRefStructIsNull(ref TestStruct? testStruct);
+
+        int GetValueInNullableRefStruct(ref TestStruct? testStruct);
+        void SetValueInNullableRefStruct(ref TestStruct? testStruct);
+    }
+}
