+++ AdvancedDLSupport.Tests/Data/Interfaces/ITypeLoweringLibrary.cs
@@ -3,7 +3,7 @@ using System.Runtime.InteropServices;
 
 namespace AdvancedDLSupport.Tests.Data
 {
-    public interface IComplexTypeLibrary
+    public interface ITypeLoweringLibrary
     {
         string GetString();
         string GetNullString();
@@ -12,10 +12,5 @@ namespace AdvancedDLSupport.Tests.Data
         bool CheckIfStringIsNull(string value);
 
         UIntPtr StringLength(string value);
-        TestStruct? GetAllocatedTestStruct();
-        TestStruct? GetNullTestStruct();
-
-        [return: MarshalAs(UnmanagedType.I1)]
-        bool CheckIfStructIsNull(TestStruct? testStruct);
     }
 }
